+++ packages/react-pose-text/src/index.ts
@@ -1,0 +1,3 @@
+import SplitText from './SplitText';
+
+export default SplitText;