+++ playground/index.js
@@ -86,6 +86,12 @@ import {
 // .add('increased stiffness', () => <SpringStiffer />)
 // .add('increased stiffness and damping', () => <SpringStifferDamping />);
 
+import { AnimateText, TextPoseGroup, ExternalPoser } from "./plugins/text";
+storiesOf("SplitText", module)
+  .add("animate text", () => <AnimateText />)
+  .add("pose group", () => <TextPoseGroup />)
+  .add("external poser", () => <ExternalPoser />);
+
 storiesOf("tween", module)
   .add("default props", () => <TweenBasic />)
   .add("loop", () => <TweenLoop />)